Fix GCC 10+ build
GCC 10 changed the default option from `-fcommon` to `-fno-common`.

A quick and dirty fix could be to just set `-fcommon` in CMakeLists.txt,
but the proper fix is to move the variables to a single compilation unit
and simply reference them in the other CU, instead of each CU having its
own copy.

Signed-off-by: Eric Engestrom <eric@engestrom.ch>